refactor(wandb): move common functions to trainer_utils
